+++ server/src/base/base.model.ts
