+++ server/src/character/character.model.ts
@@ -1,11 +1,11 @@
 import { Column } from "sequelize-typescript";
-import { Base, uniqString } from "src/base/character.model";
+import { Base, uniqString } from "src/base/base.model";
 
 interface CharacterCreationAttrs {
-  title: string;
+	title: string;
 }
 
 export class Character extends Base<Character, CharacterCreationAttrs> {
-  @Column(uniqString)
-  title: string;
+	@Column(uniqString)
+	title: string;
 }
