+++ server/src/items/items.model.ts
@@ -1,12 +1,12 @@
 import {
-  BelongsTo,
-  BelongsToMany,
-  Column,
-  DataType,
-  ForeignKey,
-  Table,
+	BelongsTo,
+	BelongsToMany,
+	Column,
+	DataType,
+	ForeignKey,
+	Table,
 } from "sequelize-typescript";
-import { Base, requireString } from "src/base/character.model";
+import { Base, requireString } from "src/base/base.model";
 import { Collection } from "src/collections/collections.model";
 import { FieldItem } from "src/fields-items/fields-items.model";
 import { Field } from "src/fields/fields.model";
@@ -14,28 +14,28 @@ import { ItemTag } from "src/items-tags/items-tags.model";
 import { Tag } from "src/tags/tags.model";
 
 interface ItemCreationAttrs {
-  title: string;
-  collectionId: number;
-  img?: string;
+	title: string;
+	collectionId: number;
+	img?: string;
 }
 
 @Table({ tableName: "items" })
 export class Item extends Base<Item, ItemCreationAttrs> {
-  @Column(requireString)
-  title: string;
+	@Column(requireString)
+	title: string;
 
-  @Column({ type: DataType.STRING })
-  img: string;
+	@Column({ type: DataType.STRING })
+	img: string;
 
-  @ForeignKey(() => Collection)
-  @Column({ field: "collectionId" })
-  collectionId: number;
-  @BelongsTo(() => Collection)
-  collection: Collection;
+	@ForeignKey(() => Collection)
+	@Column({ field: "collectionId" })
+	collectionId: number;
+	@BelongsTo(() => Collection)
+	collection: Collection;
 
-  @BelongsToMany(() => Tag, () => ItemTag)
-  tag: Tag[];
+	@BelongsToMany(() => Tag, () => ItemTag)
+	tag: Tag[];
 
-  @BelongsToMany(() => Field, () => FieldItem)
-  field: Field[];
+	@BelongsToMany(() => Field, () => FieldItem)
+	field: Field[];
 }
